+++ echovault/api_hash.go
