+++ echovault/keyspace.go
@@ -18,8 +18,8 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"github.com/echovault/echovault/constants"
 	"github.com/echovault/echovault/internal"
-	"github.com/echovault/echovault/pkg/constants"
 	"log"
 	"math/rand"
 	"runtime"
