+++ internal/config/config.go
@@ -19,8 +19,8 @@ import (
 	"errors"
 	"flag"
 	"fmt"
+	"github.com/echovault/echovault/constants"
 	"github.com/echovault/echovault/internal"
-	"github.com/echovault/echovault/pkg/constants"
 	"log"
 	"os"
 	"path"
