+++ internal/modules/acl/acl.go
@@ -20,9 +20,9 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"github.com/echovault/echovault/constants"
 	"github.com/echovault/echovault/internal"
 	"github.com/echovault/echovault/internal/config"
-	"github.com/echovault/echovault/pkg/constants"
 	"github.com/gobwas/glob"
 	"gopkg.in/yaml.v3"
 	"log"
