+++ internal/modules/acl/commands.go
@@ -18,8 +18,8 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"github.com/echovault/echovault/constants"
 	"github.com/echovault/echovault/internal"
-	"github.com/echovault/echovault/pkg/constants"
 	"gopkg.in/yaml.v3"
 	"log"
 	"os"
