+++ internal/modules/hash/key_funcs.go
@@ -16,8 +16,8 @@ package hash
 
 import (
 	"errors"
+	"github.com/echovault/echovault/constants"
 	"github.com/echovault/echovault/internal"
-	"github.com/echovault/echovault/pkg/constants"
 )
 
 func hsetKeyFunc(cmd []string) (internal.AccessKeys, error) {
