+++ test/modules/connection/commands_test.go
@@ -18,10 +18,10 @@ import (
 	"bytes"
 	"context"
 	"errors"
+	"github.com/echovault/echovault/constants"
+	"github.com/echovault/echovault/echovault"
 	"github.com/echovault/echovault/internal"
 	"github.com/echovault/echovault/internal/config"
-	"github.com/echovault/echovault/pkg/constants"
-	"github.com/echovault/echovault/pkg/echovault"
 	"github.com/tidwall/resp"
 	"net"
 	"reflect"
