+++ test/modules/hash/api_test.go
@@ -16,8 +16,8 @@ package hash
 
 import (
 	"context"
+	"github.com/echovault/echovault/echovault"
 	"github.com/echovault/echovault/internal/config"
-	"github.com/echovault/echovault/pkg/echovault"
 	"reflect"
 	"slices"
 	"testing"
