+++ test/modules/sorted_set/commands_test.go
@@ -19,11 +19,11 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"github.com/echovault/echovault/constants"
+	"github.com/echovault/echovault/echovault"
 	"github.com/echovault/echovault/internal"
 	"github.com/echovault/echovault/internal/config"
 	"github.com/echovault/echovault/internal/modules/sorted_set"
-	"github.com/echovault/echovault/pkg/constants"
-	"github.com/echovault/echovault/pkg/echovault"
 	"github.com/tidwall/resp"
 	"math"
 	"net"
