+++ test/modules/string/api_test.go
@@ -16,8 +16,8 @@ package str
 
 import (
 	"context"
+	"github.com/echovault/echovault/echovault"
 	"github.com/echovault/echovault/internal/config"
-	"github.com/echovault/echovault/pkg/echovault"
 	"testing"
 )
