+++ test/modules/string/commands_test.go
@@ -19,10 +19,10 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"github.com/echovault/echovault/constants"
+	"github.com/echovault/echovault/echovault"
 	"github.com/echovault/echovault/internal"
 	"github.com/echovault/echovault/internal/config"
-	"github.com/echovault/echovault/pkg/constants"
-	"github.com/echovault/echovault/pkg/echovault"
 	"github.com/tidwall/resp"
 	"net"
 	"reflect"
